initial commit
And it's a really big one.

I think I found the right shape for the core API, and the
Mandrill receiver even *kind of* works. Because the core
API was shaped by my ablity to get it working with the
receiver's Servant handlers, it didn't seem worth it to
make a halfway commit that wouldn't even build.